requirements.txt
pefile
pydeep
python-magic
bitstring
pymongo
